requirements.txt
binance
tweepy
PySpark
jupyter
findspark
kafka-python
